+++ assets/src/js/components/Explorer.vue
@@ -1,22 +1,8 @@
 <template>
-  <div class="explorer">
-    <explorer-search-widget 
-      :apiUrl="endpoint"
-      @reset="handleReset">
-    </explorer-search-widget>
-
-    <div class="card my-3 px-2 py-3" v-if="hasSearchResults">
-      <div class="row no-gutters">
-        <search-results v-bind:value="this.datasets"></search-results>
-      </div>
-    </div>
-  </div>
+  <router-view></router-view>
 </template>
 
 <script>
-import ExplorerSearchWidget from "./ExplorerSearchWidget.vue";
-import SearchResults from './SearchResults.vue';
-
 export default {
   props: {
     host: {
@@ -37,6 +23,6 @@ export default {
     this.$store.state.host = this.host;
     this.$store.state.port = this.port;
     this.$store.state.ssl = this.ssl;
-  }
+  },
 }
 </script>
