+++ assets/src/js/components/ExplorerSearchWidget.vue
@@ -1,50 +1,57 @@
 <template>
-  <div class="card search-widget">
-    <div class="row no-gutters flex-lg-column">
-      <div class="space-text-section col-lg-4">
-        <div class="card-body">
-          <font-awesome-layers class="fa-2x float-left mr-2">
-            <font-awesome-icon icon="circle" class="text-info" transform="up-0.5"></font-awesome-icon>
-            <font-awesome-icon icon="pencil-alt" class="text-white" transform="up-0.5 shrink-7"></font-awesome-icon>
-          </font-awesome-layers>
-          <p class="lead card-text">Draw a search area.</p>
-          <p class="small text-muted card-text">You can use any or all of the available tools;
-          overlapping areas will be unioned.</p>
+  <div>
+    <div class="card search-widget">
+      <div class="row no-gutters flex-lg-column">
+        <div class="space-text-section col-lg-4">
+          <div class="card-body">
+            <font-awesome-layers class="fa-2x float-left mr-2">
+              <font-awesome-icon icon="circle" class="text-info" transform="up-0.5"></font-awesome-icon>
+              <font-awesome-icon icon="pencil-alt" class="text-white" transform="up-0.5 shrink-7"></font-awesome-icon>
+            </font-awesome-layers>
+            <p class="lead card-text">Draw a search area.</p>
+            <p class="small text-muted card-text">You can use any or all of the available tools;
+            overlapping areas will be unioned.</p>
+          </div>
         </div>
-      </div>
-      <div class="space-map-section col-lg-8 order-lg-last">
-        <l-map></l-map>
-      </div>
-      <div class="time-section col-lg-4">
-        <div class="card-body">
-          <font-awesome-layers class="fa-2x float-left mr-2">
-            <font-awesome-icon icon="circle" class="text-info" transform="up-0.5"></font-awesome-icon>
-            <font-awesome-icon :icon="['far', 'calendar-alt']" class="text-white" transform="up-0.5 shrink-7"></font-awesome-icon>
-          </font-awesome-layers>
-          <p class="lead card-text">Select a date range.</p>
-          <p class="small text-muted card-text">Granularity controls how Plenario groups data in
-          charts. You can always download all the individual data points within the range you
-          specify.</p>
-          <time-range-form></time-range-form>
+        <div class="space-map-section col-lg-8 order-lg-last">
+          <l-map></l-map>
         </div>
-      </div>
-      <div class="action-section col-lg-4 flex-grow-1 d-flex flex-column justify-content-end">
-        <div class="card-body">
-          <div class="row hairline-gutters">
-            <div class="col-auto">
-              <button class="btn btn-block btn-danger" @click="reset">
-                <font-awesome-icon icon="undo"></font-awesome-icon>
-                <span class="d-lg-none">Reset</span>
-              </button>
-            </div>
-            <div class="col">
-              <button class="btn btn-block btn-info" @click="search">
-                <font-awesome-icon icon="search"></font-awesome-icon>
-                Search
-              </button>
+        <div class="time-section col-lg-4">
+          <div class="card-body">
+            <font-awesome-layers class="fa-2x float-left mr-2">
+              <font-awesome-icon icon="circle" class="text-info" transform="up-0.5"></font-awesome-icon>
+              <font-awesome-icon :icon="['far', 'calendar-alt']" class="text-white" transform="up-0.5 shrink-7"></font-awesome-icon>
+            </font-awesome-layers>
+            <p class="lead card-text">Select a date range.</p>
+            <p class="small text-muted card-text">Granularity controls how Plenario groups data in
+            charts. You can always download all the individual data points within the range you
+            specify.</p>
+            <time-range-form></time-range-form>
+          </div>
+        </div>
+        <div class="action-section col-lg-4 flex-grow-1 d-flex flex-column justify-content-end">
+          <div class="card-body">
+            <div class="row hairline-gutters">
+              <div class="col-auto">
+                <button class="btn btn-block btn-danger" @click="reset">
+                  <font-awesome-icon icon="undo"></font-awesome-icon>
+                  <span class="d-lg-none">Reset</span>
+                </button>
+              </div>
+              <div class="col">
+                <button class="btn btn-block btn-info" @click="search">
+                  <font-awesome-icon icon="search"></font-awesome-icon>
+                  Search
+                </button>
+              </div>
             </div>
           </div>
         </div>
+      </div>
+    </div>
+    <div class="card my-3 px-2 py-3" v-if="hasSearchResults">
+      <div class="row no-gutters">
+        <search-results v-bind:value="this.datasets"></search-results>
       </div>
     </div>
   </div>
@@ -111,23 +118,38 @@ $widget-landscape-max-height:
 <script>
 import LMap from './LMap.vue';
 import TimeRangeForm from './TimeRangeForm.vue';
+import SearchResults from './SearchResults.vue';
 
 export default {
   components: {
     LMap,
     TimeRangeForm,
   },
-  props: {
-    /**
-     * Host URL of the target backend. This can change depending on the 
-     * application context. For example, in a development application this
-     * value could be localhost:4000, in production, it could be the URL of
-     * the live server.
-     */
-    apiUrl: {
-      required: true,
-      type: String,
-    }
+  computed: {
+    ssl: function () {
+      return this.$store.state.ssl;
+    },
+
+    host: function () {
+      return this.$store.state.host;
+    },
+
+    port: function () {
+      return this.$store.state.port;
+    },
+    
+    datasets: function () {
+      return this.$store.state.datasets;
+    },
+
+    endpoint: function () {
+      var http_or_https = this.ssl ? 'https' : 'http';
+      return `${http_or_https}://${this.host}:${this.port}/api/v2/data-sets`;
+    },
+
+    hasSearchResults: function () {
+      return this.datasets.length > 0;
+    },
   },
   methods: {
     /**
@@ -136,7 +158,7 @@ export default {
      * - Needs error handling!
      */
     search: async function (_) {
-      var json = await fetch(this.apiUrl).then(function(response) {
+      var json = await fetch(this.endpoint).then(function(response) {
         return response.json();
       });
 
@@ -148,8 +170,13 @@ export default {
      * event should clear their state.
      */
     reset: function (_) {
-      this.$emit('reset', {})
+      this.$store.commit('clearQuery');
     }
   },
+  components: {
+    LMap,
+    SearchResults,
+    TimeRangeForm,
+  }
 }
 </script>
