--- assets/src/js/components/search/TimeCardHeader.vue
@@ -1,14 +1,0 @@
-<template>
-  <div>
-    <div class="card-header">
-      <i class="far fa-clock"></i>
-      &ensp;Time
-    </div>
-  </div>
-</template>
-
-<script>
-export default {
-  name: 'TimeCardHeader'
-}
-</script>
